requirements.txt
streamlit==1.32.0
pandas==2.2.0
mysql-connector-python==8.3.0
plotly==5.18.0
openpyxl==3.1.2
xlsxwriter==3.1.2